+++ PSPocs/PSPocs_Settings.ps1
@@ -6,4 +6,8 @@
 # ---------------------------------------------------------------------------
 
 # get existing literature and document libraries from general config file
-Update-PocsLibrary
+if (Test-Path -Path $PSPocs.Config){
+    Update-PocsLibrary
+} else {
+    Write-FormattedWarning -Message "Package papis may be not installed. Initialization of module is aborted." -Module $PSPocs.Name
+}